Test specific methods removed from NodeApi.
